增加播放完成跳转到剧集推荐页

+++ app/src/main/java/cn/com/films66/app/model/FilmEvents.java
@@ -44,11 +44,11 @@ public class FilmEvents implements Parcelable {
         }
     };
 
-    public int getStartTime() {
+    public long getStartTime() {
         return TimeUtils.stringToTime(start_time);
     }
 
-    public int getEndTime() {
+    public long getEndTime() {
         return TimeUtils.stringToTime(end_time);
     }
